requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
clip @ git+https://github.com/ultralytics/CLIP.git@68c3966cf14f5bd031c144cccbf1cbe904932103
contourpy==1.2.1
cycler==0.12.1
defusedxml==0.7.1
filelock==3.15.4
filetype==1.2.0
fonttools==4.53.1
fsspec==2024.6.1
ftfy==6.2.3
idna==3.7
Jinja2==3.1.4
kiwisolver==1.4.5
lapx==0.5.9.post1
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psutil==6.0.0
py-cpuinfo==9.0.0
pybboxes==0.1.6
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.40
scipy==1.14.0
seaborn==0.13.2
six==1.16.0
supervision==0.22.0
sympy==1.13.2
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.79
ultralytics-thop==2.0.5
urllib3==2.2.2
wcwidth==0.2.13
